feat: accept and make use of note template
As discussed in #12, this commit adds a feature to let users configure
their own template note.

This commit also adds documentation on how to use the template system. I
am expecting it to grow as this project evolves and thus the need will
come to support more variables, but we will see how it goes.

Should solve #12 too.