feat: if user specifies title, then use it as filename
Many times I have tried linking back issues in Markdown but since
it makes sense to use the timestamps as the directory name and since I
was using `README.md` as the default filename, it's practically
impossible to link messages this way. What I have done when needed, I
have been searching notes and then manually linking.

This uses the title flag as filename when users provide this flag.